+++ litecli/packages/special/llm.py
@@ -222,7 +222,7 @@ def handle_llm(text, cur) -> Tuple[str, Optional[str]]:
     if "-c" in parts:
         capture_output = True
         use_context = False
-    # If the parts has `pormpt` command without `-c` then use context to the prompt.
+    # If the parts has `prompt` command without `-c` then use context to the prompt.
     # \llm -m ollama prompt "Most visited urls?"
     elif "prompt" in parts:  # User might invoke prompt with an option flag in the first argument.
         capture_output = True