+++ litecli/packages/special/llm.py
@@ -163,6 +163,7 @@ $question
 
 Explain the reason for choosing each table in the SQL query you have
 written. Keep the explanation concise.
+When the user request for multiple sql queries, include each SQL query in a separate code fence.
 Finally include a sql query in a code fence such as this one:
 
 ```sql
@@ -222,7 +223,7 @@ def handle_llm(text, cur) -> Tuple[str, Optional[str]]:
     if "-c" in parts:
         capture_output = True
         use_context = False
-    # If the parts has `prompt` command without `-c` then use context to the prompt.
+    # If the parts has `pormpt` command without `-c` then use context to the prompt.
     # \llm -m ollama prompt "Most visited urls?"
     elif "prompt" in parts:  # User might invoke prompt with an option flag in the first argument.
         capture_output = True