+++ CHANGELOG.md
@@ -5,9 +5,9 @@
 
 * Make the history file location configurable. ([#206](https://github.com/dbcli/litecli/issues/206))
 
-### Internal
+### Bug Fixes
 
-* Fix typo `pormpt`to `prompt` in `special/llm.py`.
+* Enhance llm prompt to include each SQL query in a separate code fence.
 
 
 ## 1.14.4 - 2025-01-31